+++ kernels/gemm/matmul/common.cuh
@@ -14,7 +14,7 @@ void gemm_fp32_sm89_cuda(
     cudaStream_t stream);
 
 void gemm_bf16_sm89_cuda(
-    float* output,
+    __nv_bfloat16* output,
     const __nv_bfloat16* left,
     const __nv_bfloat16* right,
     int M,
